First working code
A start!
Working OLED display, LEDs (including IR), and 3 push buttons.
The navigatable menu system can be seen working. Left and right buttins scroll the menu items and the centre button lights it up.
The green LED lights when the left button is pressed.
The IR LED lights when the centre button is pressed.
The red LED lights when the right button is pressed.

Debounce code for the buttons and menu scroll speed control has also been implemented.