Structure change + Intervalometer started
A new file for button management (which is now all contained within functions). this enables buttons to be used within loops.

I've started code on the intervalometer. There's a set value for interval at the moment. Functions can display how many hours, minutes and seconds are held.

The debug bug has returned unfortunately. This will go away when less dynamic memory is used. Code will be changed in the future to store strings in flash memory which will make a big difference.