Hardware change - Bug fixes - Days
Fixed issues #1  and #2 .

Added an external pullup resistor for each button (1k Ohm) which has stopped the random behaviour). Code has been changed slightly to reflect this.

Added days as a unit of time for settings and display.

Time is now accurately saved when entered and accurately displayed.

Tweaked the flashing time for editing values so that it's easier to see.

Updated Readme to reflect changes (formatting updates will follow).

Fixed a bug which caused the setTime function to exit when the back button was pressed during editing a value (instead of just exiting the edit).