Intervalometer in a working state
Added a function to accept numbers up to 99999 from the user. This has been used to accept numbers for the intervalometer.

Added the intervalometer functionality. Users can now select an interval, a duration (none, a time or a number of shots) and start the intervalometer.

Placeholder variables have been placed for multiple trigger devices (such as cameras and flashes). These have been implemented into an early trigger() function. Calling trigger() will trigger all selected devices in sequence. There is not yet an option to set these devices and the code for the devices is not yet implemented (except for Nikon infrared).